Small style updates and breaking out components
