+++ src/pages/index.tsx
@@ -14,11 +14,8 @@ export default function Home() {
       </Head>
       <main className="flex min-h-screen flex-col items-center bg-gradient-to-b from-[#2e026d] to-[#15162c]">
         <div className="container flex flex-col items-center justify-center gap-12 px-4 py-16 ">
-          <Header variant={1}>Giftly</Header>
-          <div className="flex flex-col items-center gap-2">
-            My Wishlists
-            <WishLists />
-          </div>
+          <header className="text-5xl font-bold text-white">Giftly</header>
+          <WishLists />
         </div>
         <BottomMenu>
           <AuthShowcase />
